twitter/tweetapi/views.py
from django.shortcuts import render
from rest_framework import generics
from tweet.models import Tweet
from .serializers import TweetSerializer

class TweetListCreateAPIView(generics.ListCreateAPIView):
    queryset = Tweet.objects.all()
    serializer_class = TweetSerializer

class TweetDetailAPIView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Tweet.objects.all()
    serializer_class = TweetSerializer
